Request messages every 2s in ChatCLI. Auto resize buffer in Server.

+++ Server/src/client/client_cli.py
@@ -1,11 +1,31 @@
 import datetime
+import time
 import os
+import signal
+import sys
 from client import Client
 
 
+def message_tostr(m):
+    time_str = datetime.datetime.fromtimestamp(m.timeUnix).strftime('%H:%M:%S')
+    return '[%s] %d: %s' % (time_str, m.senderID, m.text)
+
+
+def input_intime(timeout=0):
+    def interrupted(signum, frame):
+        raise TimeoutError
+    signal.signal(signal.SIGALRM, interrupted)
+    signal.alarm(timeout)
+    try:
+        return input()
+    finally:
+        signal.alarm(0)
+
+
 class ChatCLI:
-    client = Client()
+    client = None
     targetId = None
+    last_recv_time = time.time()
 
     def __init__(self, client, targetId):
         self.client = client
@@ -26,13 +46,32 @@ class ChatCLI:
     def handle_sendfile(self, args):
         self.client.send_file(self.targetId, filepath=args[0])
 
+    def recv_messages(self):
+        msgs = self.client.get_messages(afterTime=self.last_recv_time)
+        self.last_recv_time = time.time()
+        msgs = [m for m in msgs if m.senderID == self.targetId]
+        for m in msgs:
+            print(message_tostr(m))
+
     def run(self):
+        TIMEOUT = 2
         while True:
             print('chat %d > ' % self.targetId, end='')
-            cmd = input()
+            while True:
+                try:
+                    cmd = input_intime(TIMEOUT)
+                    break
+                except TimeoutError:
+                    try:
+                        self.recv_messages()
+                    except Exception as e:
+                        print('Error: ' + str(e))
+
             tokens = cmd.split(' ')
             if tokens[0] == 'exit':
                 return
+            if cmd == '':
+                continue
             if not hasattr(self, 'handle_' + tokens[0]):
                 print('Invalid command')
                 continue
@@ -43,8 +82,11 @@ class ChatCLI:
 
 
 class ClientCLI:
-    client = Client()
-    DOWNLOAD_PATH = '~/Downloads'
+    client = None
+    DOWNLOAD_PATH = '/Users/wangrunji/Downloads/'
+
+    def __init__(self, client) -> None:
+        self.client = client
 
     def handle_help(self, args):
         HELP = """
@@ -83,23 +125,28 @@ class ClientCLI:
             if u.isFriend:
                 print(u.username)
 
+    def handle_profile(self, args):
+        info = self.client.get_self()
+        print(info)
+
     def handle_recvmsg(self, args):
         msgs = self.client.recv_message()
         for m in msgs:
             if m.text:
-                time_str = datetime.datetime.fromtimestamp(m.timeUnix).strftime('%H:%M:%S')
-                print('[%s] %d: %s' % (time_str, m.senderID, m.text))
+                print(message_tostr(m))
 
     def handle_recvfile(self, args):
         msgs = self.client.recv_message()
         for m in msgs:
             if m.file:
-                path = os.path.join(self.DOWNLOAD_PATH, 'file')
+                time_str = datetime.datetime.fromtimestamp(m.timeUnix).strftime('%Y%m%d_%H%M%S')
+                path = os.path.join(self.DOWNLOAD_PATH, 'file_'+time_str)
+                if os.path.exists(path):
+                    continue
                 file = open(path, 'wb')
                 file.write(m.file)
                 file.close()
-                print('Saved file to download path')
-
+                print('File saved to: ' + path)
 
     def handle_chat(self, args):
         userid = self.client.get_userid(username=args[0])
@@ -121,5 +168,8 @@ class ClientCLI:
 
 
 if __name__ == '__main__':
-    cli = ClientCLI()
+    host = sys.argv[1] if len(sys.argv) >= 2 else 'localhost:8000'
+    print('Linking to ' + host)
+    client = Client(host)
+    cli = ClientCLI(client)
     cli.run()
